+++ app/Tabletop/Bestiary/Exceptions/BestiaryDeploymentDenied.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GreatMarketrealmTabletop\Tabletop\Bestiary\Exceptions;
+
+use RuntimeException;
+
+defined('ABSPATH') || exit;
+
+final class BestiaryDeploymentDenied extends RuntimeException
+{
+}